+++ components/family/diary/family-diary-contents.tsx
@@ -1,6 +1,6 @@
-import { useFetchDiary } from '../../../hooks/queries/family';
-import FamilyDiaryItem from '@/components/family/diary/family-diary-item';
-import FamilyDiaryContentsEmpty from '@/components/family/diary/family-diary-empty';
+import { useFetchDiary } from '@/hooks/queries/family';
+import FamilyDiaryItem from '@/components/family/diary/item/family-diary-item';
+import FamilyDiaryContentsEmpty from '@/components/family/diary/empty/family-diary-empty';
 
 export default function FamilyDiaryContents() {
   const diaryQuery = useFetchDiary('2024-06-16');
