+++ components/family/growth/family-growth-contents.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useFetchGrowth } from '@/hooks/queries/family';
-import FamilyGrowthList from './family-growth-list';
-import FamilyGrowthEmpty from './family-growth-empty';
+import FamilyGrowthList from './list/family-growth-list';
+import FamilyGrowthEmpty from './empty/family-growth-empty';
 
 export default function FamilyGrowthContents() {
   const growthQuery = useFetchGrowth('2024-06-16');
