+++ components/family/growth/item/family-growth-item.module.scss
