+++ components/family/growth/list/family-growth-list.module.scss
