+++ components/family/schedule/family-schedule-contents.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useFetchSchedule } from '@/hooks/queries/family';
-import FamilyScheduleEmpty from './family-schedule-empty';
-import FamilyScheduleList from './family-schedule-list';
+import FamilyScheduleEmpty from './empty/family-schedule-empty';
+import FamilyScheduleList from './list/family-schedule-list';
 
 export default function FamilyScheduleContents() {
   const scheduleQueries = useFetchSchedule('2024-06-16');
