--- components/family/schedule/family-schedule-fetcher.tsx
@@ -1,10 +1,0 @@
-import React, { PropsWithChildren } from 'react';
-import { useFetchSchedule } from '@/hooks/queries/family';
-
-export default function FamilyScheduleFetcher(props: PropsWithChildren) {
-  const scheduleQueries = useFetchSchedule('2024-06-16');
-
-  if (scheduleQueries.isLoading) return <p>로딩중입니다.</p>;
-
-  return props.children;
-}
