+++ components/family/schedule/family-schedule.tsx
@@ -1,15 +1,17 @@
 import React from 'react';
 import FamilyScheduleContents from './family-schedule-contents';
-import FamilyScheduleFetcher from './family-schedule-fetcher';
 import styles from './family-schedule.module.scss';
+import Suspensive from '@/components/suspensive/suspensive';
+import { useFetchSchedule } from '@/hooks/queries/family';
 
 export default function FamilySchedule() {
+  const scheduleQuery = useFetchSchedule('2024-06-16');
   return (
     <section className={styles.wrapper}>
       <h2 className={styles.title}>일정기록</h2>
-      <FamilyScheduleFetcher>
+      <Suspensive isLoading={scheduleQuery.isLoading}>
         <FamilyScheduleContents />
-      </FamilyScheduleFetcher>
+      </Suspensive>
     </section>
   );
 }
