+++ components/family/schedule/list/family-schedule-list.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import FamilyScheduleItem from '@/components/family/schedule/family-schedule-item';
+import FamilyScheduleItem from '@/components/family/schedule/item/family-schedule-item';
 import styles from './family-schedule-list.module.scss';
 
 export default function FamilyScheduleList() {
